requirements.txt
torch
torchvision
torchmetrics
pytest
numpy
transfer-learning-vision-classifiers
pillow